Fix: Warnings related to importing numpy
Remove all MATLAB imports of `numpy`. This fixes warnings related to
numpy import conflicting with functions in the global namespace.
Warnings all fit this pattern: `[Warning: The name 'char' is already in
use as a method name.  This will become an error in a future release.]`.
All of the test cases are passing, so this seems like it works.

This also reduces the amount of noise in the test output makes it easier
for the developer/users to read the output.

The current tests seem adequate to cover the removal of these imports.
It may also be nice to wrap the mhkit/mhkit_python_utils imports into a
function to keep the code dry and add an error message if the user does
not have either of the above python libraries installed. We could also
add a check for the version of mhkit.