Tests: Add code coverage
Use built in MATLAB code coverage tool to generate code coverage report.

There are two warnings when this runs for which fixes are not readily
apparent:

Warning: Coverage result is empty for check_range.m because the file is invalid.

Warning: Coverage result is empty for blade_moments.m because the file is invalid.

These two warnings indicate that code coverage is not being generated
for those files.