Actions: Attempt to fix artifact upload/download path
Try to correctly set name and path. This sets the path and name on
upload, but only sets the name on download. We need the
mhkit_webread_cache directory to be in the root directory of this
project and before this commit it was in <root
directory>/mhkit_webread_cache/mhkit_webread_cache