D3D: Add delft_3d_get_keys
Key extraction for a netCDF4 object in Python is natively supported.
Within MATLAB nested key extraction is much more difficult due to nested
structures. With additional effort this could be written in MATLAB.

In this code we extract the keys and values in Python and create two different lists that we pass to MATLAB. The lists are parsed into a struct that is returned to the user.