D3D: Add delft_3d_get_all_data_points function
This function introduces combined data structures for this module.

Some MHKiT-Python D3D functions return dataframes. Within this module we
use the convert_numeric_dataframe_to_struct function to convert a pandas
dataframe to a struct. We then add the dataframe to the struct using the
`df` fieldname. This maintains the original output dataframe, and the
converts the data into a more usable format.