Fix: Properly convert doubles with multiple values into numpy arrays
In MATLAB a double can be a single value or an array. If the double has
multiple values, to pass into python we must convert to an array. The
previously version of this code naively converted all doubles into a
numpy array, inadvertently converting singlar values into numpy arrays.
In MHKiT-Python v0.8.0 using mhkit.river.io.d3d.create_points with
numpy arrays of length 1 yields the following error:

```
TypeError: len() of unsized object
```

The fix in MHKiT-MATLAB is to not create numpy arrays of one value

This also seems like we exposed a potential bug in MHKiT-Python v0.8.0.