Fix: Correctly create an above/below surface mask
The previous code here appeared to use an xarray features that allow for
comparison across coordinates/dimensions. This more explicitly creates a
matrix of ranges over time and outputs bds as a binary matrix with 1
denoting a value below the surface interference distance and 0 a value
above the surface interference distance.